Work on consolidating server adapter and clearing house for simplicity.
